Add MODE environment variable
When running tests on travis, NODE_ENV is set to test
same as when running tests locally. Issue is that the DATABASE_URL
used provided by heroku requires to receive ssl property of true,
which is not required in the local testing environment.
So, to differentiate the two similar environments, I use environment
variable MODE=online to detect that the test is running online on
travis. it could be any other CI tool.